Extract assertSelectedCells() common test util function.
